ctf/web/DoxPit/challenge/av/requirements.txt
bcrypt==4.1.2
Faker==25.2.0
Flask==3.0.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
